Refactored code to store Gomoku  state
Needed to store Gomoku state to be used in AI
