+++ src/GomokuAI/Interfaces/IPlayer.cs
@@ -1,9 +1,9 @@
 ﻿using GomokuAI.Engine;
 
-namespace GomokuAI.Players;
+namespace GomokuAI.Interfaces;
 
 public interface IPlayer
 {
-    public (int row, int column) GetMove();
+    public (int row, int column) GetMove(Gomoku gomoku);
     public void PrintMove(int row, int column);
 }
